[IMP] delivery_schenker_picking_volume: Adding unittests
