+++ delivery_schenker_picking_volume/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_delivery_schenker
